Remove string concats

+++ database/tmdb/tmdb_movie.sql
@@ -3,7 +3,7 @@ CREATE TABLE tmdb_movie (
 adult                     BOOLEAN NOT NULL,
 backdrop_path             TEXT,
 belongs_to_collection     JSON, -- how to handle this object?
-budget                    INTEGER NOT NULL,
+budget                    BIGINT NOT NULL,
 genres                    JSON NOT NULL, -- this could just be array of genre INTEGERs
 homepage                  TEXT,
 id                        INTEGER NOT NULL UNIQUE,
@@ -16,7 +16,7 @@ poster_path               TEXT,
 production_companies      JSON NOT NULL, -- how to handle this object?
 production_countries      JSON NOT NULL, -- how to handle this object?
 release_date              TEXT NOT NULL,
-revenue                   INTEGER NOT NULL,
+revenue                   BIGINT NOT NULL,
 runtime                   INTEGER,
 spoken_languages          JSON NOT NULL, -- how to handle this object?
 status                    TEXT NOT NULL,
@@ -24,7 +24,7 @@ tagline                   TEXT,
 title                     TEXT NOT NULL,
 video                     BOOLEAN NOT NULL,
 vote_average              DECIMAL NOT NULL,
-vote_count                INTEGER NOT NULL,
+vote_count                BIGINT NOT NULL,
 -- custom fields
 created_at                TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP, 
 updated_at                TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
